:recycle: redisConfig 재설정
